+++ src/Platform/Menu.php
@@ -114,11 +114,11 @@ class Menu
         return $this->findAllChildren($location)
             ->sortBy('sort')
             ->filter(function ($value) {
-                if (! $value['childs']) {
-                    return true;
+                if ($value['childs'] && $value['hideEmpty']) {
+                    return $this->showCountElement($value['slug']);
                 }
 
-                return $this->showCountElement($value['slug']);
+                return true;
             })
             ->map(static function ($value) use ($template) {
                 return view($template, $value)->render();